WIP refactored Camera component
