+++ src/components/Camera/index.js
@@ -1,199 +1,17 @@
 // @flow
 import * as React from 'react'
 import { h } from 'preact'
-import Webcam from 'react-webcam-onfido'
-import Dropzone from 'react-dropzone'
-import Visibility from 'visibilityjs'
-import classNames from 'classnames'
+import AutoCapture from './AutoCapture'
+import Photo from './Photo'
+import Video from './Video'
 
-import {cloneCanvas} from '../utils/canvas.js'
-import { asyncFunc } from '../utils/func'
-import { Overlay } from '../Overlay'
-import Title from '../Title'
+import type { CameraType } from './CameraTypes'
 
-import theme from '../Theme/style.css'
-import style from './style.css'
-
-
-const UploadFallback = ({onUploadFallback, onFallbackClick, method, i18n}) =>
-  <Dropzone
-    onDrop={([file]) => onUploadFallback(file)}
-    className={style.uploadFallback}
-    multiple={false}>
-    <button onClick={onFallbackClick()}>{i18n.t(`capture.${method}.help`)}</button>
-  </Dropzone>
-
-const CaptureActions = ({handleClick, btnText, isFullScreen, recording, liveness}) =>
-  <div className={style.captureActions}>
-    <button
-      className={classNames(
-        theme.btn, theme["btn-centered"],
-        theme["btn-primary"],
-        { [style.fullScreenBtn]: isFullScreen && !liveness, [style.recording]: recording, [style.liveness]: liveness }
-      )}
-      onClick={handleClick}>
-      <div className={classNames({[style.btnText]: isFullScreen})}>{btnText}</div>
-    </button>
-  </div>
-
-type CameraCommonType = {
-  autoCapture: boolean,
-  method: string,
-  title: string,
-  subTitle: string,
-  onUserMedia: Function,
-  onUploadFallback: File => void,
-  onWebcamError: Function,
-  onUserMedia: void => void,
-  i18n: Object,
-  isFullScreen: boolean
-}
-
-type CameraPureType = {
-  ...CameraCommonType,
-  onFallbackClick: void => void,
-  faceCaptureClick: void => void,
-  handleVideoClick: void => void,
-  webcamRef: React.Ref<typeof Webcam>,
-}
-
-// Specify just a camera height (no width) because on safari if you specify both
-// height and width you will hit an OverconstrainedError if the camera does not
-// support the precise resolution.
-const CameraPure = ({method, autoCapture, title, subTitle, onUploadFallback, onFallbackClick,
-  onUserMedia, handleClick, btnText, webcamRef, isFullScreen, onWebcamError, i18n, liveness, recording}: CameraPureType) => {
-    return (
-      <div className={style.camera}>
-        <Title {...{title, subTitle, isFullScreen}} smaller={true}/>
-        <div className={classNames(style["video-overlay"], {[style.overlayFullScreen]: isFullScreen})}>
-          <Webcam
-            className={style.video}
-            audio={!!liveness}
-            height={720}
-            {...{onUserMedia, ref: webcamRef, onFailure: onWebcamError}}
-          />
-          <Overlay {...{method, isFullScreen}}/>
-          <UploadFallback {...{onUploadFallback, onFallbackClick, method, i18n}}/>
-        </div>
-        { !autoCapture && <CaptureActions {...{isFullScreen, handleClick, btnText, liveness, recording}}/> }
-      </div>
-    )
-  }
-
-
-
-type CameraType = {
-  ...CameraCommonType,
-  onScreenshot: Function,
-  trackScreen: Function,
-}
-
-type CameraState = {
-  recording: boolean
-}
-
-//
-// class AutoCapture extends React.Component {
-//
-// }
-//
-// class Photo extends React.Component {
-//
-// }
-//
-// class Video extends React.Component {
-
-// }
-
-export default class Camera extends React.Component<CameraType, CameraState> {
-
-  webcam: ?React$ElementRef<typeof Webcam> = null
-  interval: ?Visibility
-
-  capture = {
-    start: () => {
-      this.capture.stop()
-      this.interval = Visibility.every(1000, this.screenshot)
-    },
-    stop: () => Visibility.stop(this.interval),
-    once: () => {
-      this.screenshot()
-    }
-  }
-
-  video = {
-    start: () => {
-      this.webcam && this.webcam.startRecording()
-      this.setState({recording: true})
-    },
-    stop: () => {
-      this.webcam && this.webcam.stopRecording()
-      this.setState({recording: false})
-    }
-  }
-
-  webcamMounted () {
-    const { autoCapture } = this.props
-    if (autoCapture) this.capture.start()
-  }
-
-  webcamUnmounted () {
-    this.capture.stop()
-  }
-
+export default class Camera extends React.Component<CameraType> {
   componentDidMount () {
-    this.webcamMounted()
     this.props.trackScreen('camera')
   }
-
-  componentWillUnmount () {
-    this.webcamUnmounted()
-  }
-
-  screenshot = () => {
-    const { onScreenshot } = this.props
-    const canvas = this.webcam && this.webcam.getCanvas()
-    if (!canvas){
-      console.error('webcam canvas is null')
-      return
-    }
-    asyncFunc(cloneCanvas, [canvas], onScreenshot)
-  }
-
-  handleVideoStop = () => {
-    this.video.stop()
-    const blob = this.webcam.getVideoBlob()
-    this.props.onVideoRecorded(blob)
-  }
-
-  handleVideoClick = () => {
-    this.state.recording ? this.handleVideoStop() : this.video.start()
-  }
-
-  stopCamera = () => {
-    this.capture.stop()
-  }
-
-  btnText = () => {
-    const { liveness, i18n } = this.props
-    if (!liveness) return i18n.t('capture.face.button')
-    return this.state.recording ? i18n.t('capture.liveness.stop') : i18n.t('capture.liveness.start')
-  }
-
-  handleClick = () => this.props.liveness ? this.handleVideoClick() : this.capture.once()
-
-  render() {
-    const btnText = this.btnText()
-    return (
-      <CameraPure {...{
-        ...this.props,
-        handleClick: this.handleClick,
-        webcamRef: (c) => { this.webcam = c },
-        onFallbackClick: () => { this.stopCamera },
-        recording: this.state.recording,
-        btnText
-      }}
-      />
-    )
-  }
+  render = () =>
+    this.props.autoCapture ? <AutoCapture {...this.props} /> :
+      this.props.liveness ? <Video {...this.props} /> : <Photo {...this.props} />
 }
